+++ shuup/xtheme/admin_module/__init__.py
@@ -13,12 +13,13 @@ from django_jinja.backend import Jinja2
 
 from shuup.admin.base import AdminModule, MenuEntry, Notification
 from shuup.admin.menu import CONTENT_MENU_CATEGORY
+from shuup.admin.utils.permissions import get_default_model_permissions
 from shuup.admin.utils.urls import admin_url, derive_model_url, get_edit_and_list_urls
 from shuup.admin.views.home import HelpBlockCategory, SimpleHelpBlock
 from shuup.utils.django_compat import reverse
 from shuup.xtheme._theme import get_current_theme
 from shuup.xtheme.engine import XthemeEnvironment
-from shuup.xtheme.models import Snippet
+from shuup.xtheme.models import Font, Snippet
 
 
 class XthemeAdminModule(AdminModule):
@@ -44,15 +45,20 @@ class XthemeAdminModule(AdminModule):
                 name="xtheme.config_detail",
             ),
             admin_url(
+                r"^xtheme/font/$",
+                "shuup.xtheme.admin_module.views.FontListView",
+                name="xtheme.font.list",
+            ),
+            admin_url(
                 r"^xtheme/font/new/$",
                 "shuup.xtheme.admin_module.views.FontEditView",
-                name="xtheme.admin_font.new",
+                name="xtheme.font.new",
                 kwargs={"pk": None},
             ),
             admin_url(
                 r"^xtheme/font/(?P<pk>\d+)/$",
                 "shuup.xtheme.admin_module.views.FontEditView",
-                name="xtheme.admin_font.edit",
+                name="xtheme.font.edit",
             ),
             admin_url(
                 r"^xtheme/admin-configure/",
@@ -70,7 +76,15 @@ class XthemeAdminModule(AdminModule):
                 url="shuup_admin:xtheme.config",
                 category=CONTENT_MENU_CATEGORY,
                 ordering=1,
-            )
+            ),
+            MenuEntry(
+                text=_("Fonts"),
+                icon="fa fa-font",
+                url="shuup_admin:xtheme.font.list",
+                category=CONTENT_MENU_CATEGORY,
+                subcategory="design",
+                ordering=10,
+            ),
         ]
 
     def get_help_blocks(self, request, kind):
@@ -107,6 +121,12 @@ class XthemeAdminModule(AdminModule):
                         title=_("Theming"),
                         url="shuup_admin:xtheme.config",
                     )
+
+    def get_model_url(self, object, kind, shop=None):
+        return derive_model_url(Font, "shuup_admin:xtheme.font", object, kind)
+
+    def get_required_permissions(self):
+        return get_default_model_permissions(Font)
 
 
 class XthemeSnippetsAdminModule(AdminModule):