+++ CHANGELOG.md
@@ -24,6 +24,7 @@ List all changes after the last release here (newer on top). Each change on a se
 - Admin: Migrate primary, secondary, success, danger, and text color SCSS variables to CSS variables
 - Admin: Made active state side menu list icon share same color as text
 - Admin: Made the CSS variables stem from initialized Sass variables in Colors object
+- Admin: Deprecated and migrated Shuup Typography into Xtheme
 
 ### Added